Reorganize to allow c++ in internal headers
