--- include/fenix_data_policy.h
@@ -1,67 +1,0 @@
-/*
-//@HEADER
-// ************************************************************************
-//
-//
-//            _|_|_|_|  _|_|_|_|  _|      _|  _|_|_|  _|      _|
-//            _|        _|        _|_|    _|    _|      _|  _|
-//            _|_|_|    _|_|_|    _|  _|  _|    _|        _|
-//            _|        _|        _|    _|_|    _|      _|  _|
-//            _|        _|_|_|_|  _|      _|  _|_|_|  _|      _|
-//
-//
-//
-//
-// Copyright (C) 2016 Rutgers University and Sandia Corporation
-//
-// Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
-// the U.S. Government retains certain rights in this software.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-//
-// 1. Redistributions of source code must retain the above copyright
-// notice, this list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright
-// notice, this list of conditions and the following disclaimer in the
-// documentation and/or other materials provided with the distribution.
-//
-// 3. Neither the name of the Corporation nor the names of the
-// contributors may be used to endorse or promote products derived from
-// this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY SANDIA CORPORATION "AS IS" AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
-// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SANDIA CORPORATION OR THE
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-// Author Marc Gamell, Eric Valenzuela, Keita Teranishi, Manish Parashar,
-//        Michael Heroux, and Matthew Whitlock
-//
-// Questions? Contact Keita Teranishi (knteran@sandia.gov) and
-//                    Marc Gamell (mgamell@cac.rutgers.edu)
-//
-// ************************************************************************
-//@HEADER
-*/
-
-#ifndef __FENIX_DATA_POLICY_H__
-#define __FENIX_DATA_POLICY_H__
-
-#include <mpi.h>
-#include "fenix.h"
-#include "fenix_data_group.h"
-
-int __fenix_policy_get_group(fenix_group_t** group, MPI_Comm comm, 
-      int timestart, int depth, int policy_name, void* policy_value, int* flag);
-
-#endif //__FENIX_DATA_POLICY_H__
