+++ include/fenix_init.h
@@ -44,7 +44,7 @@
 // NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-// Author Marc Gamell, Eric Valenzuela, Keita Teranishi, Manish Parashar
+// Author Marc Gamell, Eric Valenzuela, Keita Teranishi, Manish Parashar,
 //        Rob Van der Wijngaart, Michael Heroux, and Matthew Whitlock
 //
 // Questions? Contact Keita Teranishi (knteran@sandia.gov) and
@@ -53,23 +53,24 @@
 // ************************************************************************
 //@HEADER
 */
-#ifndef __FENIX_PROCES_RECOVERY_GLOBAL_H__
-#define __FENIX_PROCES_RECOVERY_GLOBAL_H__
+
+#ifndef __FENIX_INIT__
+#define __FENIX_INIT__
 
 #include <mpi.h>
 #include <setjmp.h>
-#include <stdio.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include <stdarg.h>
-#include <stdint.h>
-#include <signal.h>
 
-#include "fenix_opt.h"
-#include "fenix_util.h"
-#include "fenix_data_group.h"
+#if defined(c_plusplus) || defined(__cplusplus)
+extern "C" {
+#endif
+
+int __fenix_preinit(int *, MPI_Comm, MPI_Comm *, int *, char ***, int, int, MPI_Info, int *, jmp_buf *);
 
 
-/* This header file is intended to provide global variable definitions for fenix_process_recovery.c only */
+void __fenix_postinit(int *);
 
-#endif // __FENIX_PROCES_RECOVERY_GLOBAL_H__
+#if defined(c_plusplus) || defined(__cplusplus)
+}
+#endif
+
+#endif
