+++ include/fenix_process_recovery.hpp
@@ -54,85 +54,74 @@
 //@HEADER
 */
 
-#include <stdlib.h>
+#ifndef __FENIX_PROCESS_RECOVERY__
+#define __FENIX_PROCESS_RECOVERY__
+
+#include <mpi.h>
+#include <setjmp.h>
 #include <stdio.h>
-#include <fenix.h>
-#include <fenix_process_recovery.h>
-#include <mpi-ext.h>
+#include <unistd.h>
+#include <stdlib.h>
+#include <stdarg.h>
+#include <stdint.h>
+#include <signal.h>
 
-fenix_comm_list_t my_list = {NULL, NULL};
+#include "fenix_init.h"
 
-int __fenix_comm_push(MPI_Comm *comm) {
-  fenix_comm_list_elm_t *current = (fenix_comm_list_elm_t *) malloc(sizeof(fenix_comm_list_elm_t));
-  if (!current) return 0;
-  current->next = NULL;
-  current->comm = comm;
-  if (!my_list.tail) {
-    /* if list was empty, initialize head and tail                             */
-    current->prev = NULL;
-    my_list.head = my_list.tail = current;
-  }
-  else {
-    /* if list was not empty, add element to the head of the list              */
-    current->prev = my_list.head;
-    my_list.head->next = current;
-    my_list.head = current;
-  }
-  return FENIX_SUCCESS;
-}
+#define __FENIX_RESUME_AT_INIT 0 
+#define __FENIX_RESUME_NO_JUMP 200
 
-int __fenix_comm_delete(MPI_Comm *comm) {
+typedef void (*recover)( MPI_Comm, int, void *);
 
-  fenix_comm_list_elm_t *current = my_list.tail;
-  while (current) {
-    if (*(current->comm) == *comm) {
-      if (current != my_list.head && current != my_list.tail) {
-	current->prev->next = current->next;
-        current->next->prev = current->prev;
-      }
-      else if (current == my_list.tail) {
-        if (current->next) {
-          current->next->prev = NULL;
-          my_list.tail = current->next;
-	}
-        else my_list.tail = my_list.head = NULL;
-      }
-      else {
-        if (current->prev) {
-          current->prev->next = NULL;
-          my_list.head = current->prev;
-	}
-        else my_list.tail = my_list.head = NULL;
-      }
-      MPIX_Comm_revoke(*comm);
-      PMPI_Comm_free(comm);
-      free(current);
-      return 1;
-    }
-    else current = current->next;
-  }
-  /* if we end up here, the requested communicator has not been found */
-  return 0;
-}
-  
+typedef struct fcouple {
+    recover x;
+    void *y;
+} fenix_callback_func;
 
-void __fenix_comm_list_destroy(void) {
-  if (my_list.tail == NULL) {
-    return;
-  }
-  else {
-    fenix_comm_list_elm_t *current = my_list.tail;
-    while (current->next) {
-      fenix_comm_list_elm_t *next = current->next;
-      MPIX_Comm_revoke(*current->comm);
-      PMPI_Comm_free(current->comm);
-      free(current);
-      current = next;
-    }
-    MPIX_Comm_revoke(*current->comm);
-    PMPI_Comm_free(current->comm);
-    free(current);
-  }
-  my_list.tail = my_list.head = NULL;
-}
+typedef struct __fenix_callback_list {
+    fenix_callback_func *callback;
+    struct __fenix_callback_list *next;
+} fenix_callback_list_t;
 
+typedef struct __fenix_comm_list_elm {
+  struct __fenix_comm_list_elm *next;
+  struct __fenix_comm_list_elm *prev;
+  MPI_Comm *comm;
+} fenix_comm_list_elm_t;
+
+typedef struct {
+  fenix_comm_list_elm_t *head;
+  fenix_comm_list_elm_t *tail;
+} fenix_comm_list_t;
+
+int __fenix_create_new_world();
+
+int __fenix_repair_ranks();
+
+int __fenix_callback_register(void (*recover)(MPI_Comm, int, void *), void *);
+
+int __fenix_callback_pop();
+
+void __fenix_callback_push(fenix_callback_list_t **, fenix_callback_func *);
+
+void __fenix_callback_invoke_all(int error);
+
+int __fenix_callback_destroy(fenix_callback_list_t *callback_list);
+
+int* __fenix_get_fail_ranks(int *, int, int);
+
+int __fenix_spare_rank();
+
+int __fenix_get_rank_role();
+
+void __fenix_set_rank_role(int FenixRankRole);
+
+int __fenix_detect_failures(int do_recovery);
+
+void __fenix_finalize();
+
+void __fenix_finalize_spare();
+
+void __fenix_test_MPI(MPI_Comm*, int*, ...);
+
+#endif
