+++ src/fenix.cpp
@@ -54,10 +54,10 @@
 //@HEADER
 */
 
-#include "fenix_data_recovery.h"
-#include "fenix_process_recovery.h"
-#include "fenix_util.h"
-#include "fenix_ext.h"
+#include "fenix_data_recovery.hpp"
+#include "fenix_process_recovery.hpp"
+#include "fenix_util.hpp"
+#include "fenix_ext.hpp"
 #include "fenix.h"
 
 const Fenix_Data_subset  FENIX_DATA_SUBSET_FULL = {0, NULL, NULL, NULL, 0, __FENIX_SUBSET_FULL};
