+++ src/fenix_callbacks.cpp
@@ -56,13 +56,12 @@
 
 #include <assert.h>
 
-#include "fenix_comm_list.h"
-#include "fenix_ext.h"
-#include "fenix_process_recovery.h"
-#include "fenix_data_group.h"
-#include "fenix_data_recovery.h"
-#include "fenix_opt.h"
-#include "fenix_util.h"
+#include "fenix_ext.hpp"
+#include "fenix_process_recovery.hpp"
+#include "fenix_data_group.hpp"
+#include "fenix_data_recovery.hpp"
+#include "fenix_opt.hpp"
+#include "fenix_util.hpp"
 #include <mpi.h>
