+++ src/fenix_util.cpp
@@ -54,9 +54,9 @@
 //@HEADER
 */
 
-#include "fenix_opt.h"
-#include "fenix_process_recovery.h"
-#include "fenix_util.h"
+#include "fenix_opt.hpp"
+#include "fenix_process_recovery.hpp"
+#include "fenix_util.hpp"
 
 char* logname;
