+++ test/failed_spares/fenix_failed_spares.c
@@ -61,6 +61,7 @@
 #include <sys/types.h>
 #include <unistd.h>
 #include <pthread.h>
+#include <stdlib.h>
 
 const int kKillID = 1;
