+++ test/request_cancelled/fenix_req_cancelled_test.c
@@ -60,6 +60,7 @@
 #include <signal.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include <stdlib.h>
 
 const int kKillID = 1;
